# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Travel Archive API
  version: 0.1.0
  description: Private by default. JWT issuer and audience are verified; every
    private resource checks ownership. Guest photos upload separately after
    idempotent metadata import.
paths:
...
  /v1/auth/save-intent:
    post:
      operationId: createSaveIntent
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  minLength: 1
                  maxLength: 2048
                importKey:
                  type: string
                  format: uuid
              required:
                - token
                - importKey
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  grant:
                    type: string
                required:
                  - grant
                additionalProperties: false
        "400":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/public/{slug}:
    get:
      operationId: publicProfile
      security: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfile"
        "400":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/me:
    get:
      operationId: getMe
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  displayName:
                    type: string
                  deleting:
                    type: boolean
                required:
                  - id
                  - displayName
                  - deleting
                additionalProperties: false
        "400":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateMe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName:
                  type: string
                  minLength: 1
                  maxLength: 80
              required:
                - displayName
              additionalProperties: false
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  displayName:
                    type: string
                required:
                  - displayName
                additionalProperties: false
        "400":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteMe
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Success
        "400":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/memories/{visitId}:
    patch:
      operationId: updateMemory
      security:
        - bearerAuth: []
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                required:
                  - id
                additionalProperties: false
        "400":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/trips/{id}:
    delete:
      operationId: deleteTrip
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Success
        "400":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
      additionalProperties: false
...
    PublicProfile:
      type: object
      properties:
        displayName:
          type: string
        countries:
          type: array
          items:
            type: string
        stats:
          type: object
          properties:
            visits:
              type: integer
            photos:
              type: integer
          required:
            - visits
            - photos
          additionalProperties: false
      required:
        - displayName
        - countries
        - stats
      additionalProperties: false
...
    MemoryInput:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        memo:
          type: string
          maxLength: 500
      required:
        - title
        - memo
      additionalProperties: false
...